pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.0",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
]
